pyproject.toml
[project]
name = "use-env"
version = "1.0.0"
description = "Environment file processor with extensible secret providers"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]
keywords = ["cli", "environment", "secrets", "devops", "configuration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.8",
]

# Azure Key Vault provider
azure = [
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.8.0",
]

# AWS Secrets Manager provider
aws = [
    "boto3>=1.34.0",
]

# Google Cloud Secret Manager provider
gcp = [
    "google-cloud-secret-manager>=2.20.0",
]

# HashiCorp Vault provider
vault = [
    "hvac>=2.1.0",
]

# 1Password Connect provider
1password = [
    "aiohttp>=3.9.0",
]

# All cloud providers
all = [
    "use-env[azure]",
    "use-env[aws]",
    "use-env[gcp]",
    "use-env[vault]",
    "use-env[1password]",
]

[project.scripts]
use-env = "use_env.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "-q",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
